# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cuidador, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ContratoItem/detalleVistaCliente/{idContrato}": {
      "get": {
        "tags": [
          "ContratoItem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idContrato",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ContratoItem/listarContrato/{idUsuario}/{tipousuarioid}": {
      "get": {
        "tags": [
          "ContratoItem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idUsuario",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "tipousuarioid",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ContratoItem/guardarContrato": {
      "post": {
        "tags": [
          "ContratoItem"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterContractDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ContratoItemDTO": {
      "type": "object",
      "properties": {
        "observaciones": {
          "type": "string"
        },
        "horario_inicio_propuesto": {
          "format": "date-time",
          "type": "string"
        },
        "horario_fin_propuesto": {
          "format": "date-time",
          "type": "string"
        },
        "tareas_contrato": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TareasContratoDTO"
          }
        }
      },
      "additionalProperties": false
    },
...
    "RegisterContractDTO": {
      "type": "object",
      "properties": {
        "persona_cuidador_id": {
          "format": "int32",
          "type": "integer"
        },
        "persona_cliente_id": {
          "format": "int32",
          "type": "integer"
        },
        "contrato_item": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContratoItemDTO"
          }
        }
      },
      "additionalProperties": false
    },
...
    "TareasContratoDTO": {
      "type": "object",
      "properties": {
        "idTarea": {
          "format": "int32",
          "type": "integer"
        },
        "tituloTarea": {
          "type": "string"
        },
        "descripcionTarea": {
          "type": "string"
        },
        "tipoTarea": {
          "type": "string"
        },
        "estatusId": {
          "format": "int32",
          "type": "integer"
        },
        "fechaARealizar": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}